backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
pydantic==1.10.12
pyMuPDF
python-docx
transformers==4.39.3
peft==0.10.0
datasets
accelerate
torch>=2.1
scipy
sentencepiece
python-pptx
